% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artist_top_tracks.R
\name{get_artist_top_tracks}
\alias{get_artist_top_tracks}
\title{Gets top tracks from an artist}
\usage{
get_artist_top_tracks(
  artist_id,
  country,
  output = c("tidy", "raw"),
  limit = 20,
  offset = 0,
  token = my_token
)
}
\arguments{
\item{artist_id}{Required. Get Spotify Catalog information for the top tracks of this artist_id in the given country.}

\item{country}{Required. ISO-code of country.}

\item{output}{Type of output to return from the request. Default: tidy.}

\item{limit}{Maximum number of results to return. Should be between 1 and 50. Default: 20.}

\item{offset}{The index of the first result to return. Default: 0.20.}

\item{token}{A valid access token from the Spotify Accounts service: see
\url{https://developer.spotify.com/documentation/general/guides/authorization-guide/} for details. Default: my_token.}
}
\value{
Default: returns a tidy dataframe with a selection of the response. Alternatively, when output is set to raw, it returns the raw output from the
request.
}
\description{
Connects with the Spotify API and returns output from the href \url{https://api.spotify.com/v1/artists/{id}/top-tracks/}
}
