% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spotify_api_token.R
\name{get_spotify_api_token}
\alias{get_spotify_api_token}
\title{Gets a Spotify API token}
\usage{
get_spotify_api_token(client_id, client_secret)
}
\arguments{
\item{client_id}{Your client id. Required.}

\item{client_secret}{Your client secret. Required.}
}
\value{
In case the request succeeded, a token to connect with the Spotify API. On failure, a message indicating that authentication failed.
}
\description{
Gets an OAuth2.0 token to connect with the Spotify API.
}
\seealso{
The vignette Connecting-with-the-Spotify-API
}
